+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This change
 - Separate out copied code ([#53](https://github.com/mfikes/cljs-bean/issues/53))
 
 ### Fixed
+- `TransientArrayVector` should be private ([#51](https://github.com/mfikes/cljs-bean/issues/51))
 - `set-empty-colls!` doesn't need extra private meta ([#52](https://github.com/mfikes/cljs-bean/issues/52))
 
 ## [1.1.0] - 2019-06-29