+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This change
 ## [Unreleased]
 ### Changed
 - Separate out copied code ([#53](https://github.com/mfikes/cljs-bean/issues/53))
+- Simplify unwrapping code ([#59](https://github.com/mfikes/cljs-bean/issues/59))
 
 ### Fixed
 - `TransientArrayVector` should be private ([#51](https://github.com/mfikes/cljs-bean/issues/51))